requirements.txt
djangorestframework==3.11.2
Django==3.1.8
django-cors-headers
mock~=4.0.3
